// src/models/player.ts
export interface Player {
  card15: number | null;
  card16: number | null;
  columns: string[];
  hcp: number | null;
  lastUpdate: string | null;
  name: string;
  origin: string;
  points: number | null;
  pointsLost: number | null;
  changes: number | null;
  position: number | null;
  tournaments: number | null;
}